+++ x/group/keeper/msg_server.go
@@ -23,8 +23,7 @@ var _ group.MsgServer = Keeper{}
 const (
 	// TODO: Revisit this once we have propoer gas fee framework.
 	// Tracking issues https://github.com/cosmos/cosmos-sdk/issues/9054, https://github.com/cosmos/cosmos-sdk/discussions/9072
-	gasCostPerIteration   = uint64(20)
-	defaultMaxMetadataLen = uint64(255)
+	gasCostPerIteration = uint64(20)
 )
 
 func (k Keeper) CreateGroup(goCtx context.Context, req *group.MsgCreateGroup) (*group.MsgCreateGroupResponse, error) {
@@ -778,7 +777,7 @@ func (k Keeper) doAuthenticated(ctx sdk.Context, req authNGroupReq, action actio
 // assertMetadataLength returns an error if given metadata length
 // is greater than a fixed maxMetadataLen.
 func (k Keeper) assertMetadataLength(metadata []byte, description string) error {
-	if metadata != nil && uint64(len(metadata)) > k.maxMetadataLen {
+	if metadata != nil && uint64(len(metadata)) > k.config.MaxMetadataLen {
 		return sdkerrors.Wrapf(errors.ErrMaxLimit, description)
 	}
 	return nil